Archive output to S3 when no longer in-progress
Avoid the complexity of Redis and websocket-based rendering for commands that
are no longer running. After a command has been stopped for 10 minutes, move its
content out to S3 and mark it as archived. If rendering an archived command,
fetch its content from S3 and include it in the HTML directly.

We deem a command to be stale after 10 minutes instead of archiving immediately
on the running:false update because some clients seem incapable of posting all
output before marking the command as stopped.