Give clients time to catch up
If a command is marked as stopped before its output has been fully sent to the
server, and open clients will be closed before getting that output. New clients
will see the full history since we are still accepting it.

This change gives those open clients 10 minutes after a command is stopped
before the stream is closed.