Support PATCH /commands/:token
PUT is meant to replace a resource and requires all attributes be specified.
PATCH is different in that attributes may be omitted and will be left alone.
This is the semantics of the current PUT handler, so let's call it PATCH.

Note: we still support PUT and handle it like PATCH