Remove dependency on Handler from Storage
- Quantify Storage over any (MonadIO m)
- Hoist unsafeRunStorage and get404 out to Foundation, still available
  everywhere, but more closely coupled to Handler itself

This makes the test suite a bit cleaner (no use of the unsafe handler
function) and paves the way for a worker process that doesn't live
itself have to live in Handler (IO will be enough).

Moving unsafeRunStorage out another layer of abstraction is also good as
the only reason this "unsafe" function is safe where it's used is that,
in the context of a Handler, an unchecked exception is actually the
desired result (i.e. a 500 and error message) when a storage operation
fails.

Moving get404 also makes sense as having the Storage layer know how to
render a 404, while convenient, is not an elegant pattern to copy from
Persistent IMHO.