Avoid scrolling to bottom on empty output
We send an empty string over the websocket connection so we can receive
an acknowledgement from the client even if there's no actual output to
display. This change removes the scroll-to-bottom logic in this case as
that's very annoying to a user trying to scroll back up after the
command's finished.